+++ TabletopTweaks/Bugfixes/UI/FeatureSelection.cs
@@ -7,7 +7,7 @@ using System.Linq;
 
 namespace TabletopTweaks.Bugfixes.UI {
     static class FeatureSelection {
-        [HarmonyPatch(typeof(NestedFeatureSelectionUtils), "AllNestedFeaturesUnavailable", new[] { typeof(UnitDescriptor), typeof(LevelUpState), typeof(FeatureSelectionState), typeof(BlueprintFeature) })]
+        //[HarmonyPatch(typeof(NestedFeatureSelectionUtils), "AllNestedFeaturesUnavailable", new[] { typeof(UnitDescriptor), typeof(LevelUpState), typeof(FeatureSelectionState), typeof(BlueprintFeature) })]
         static class NestedFeatureSelectionUtils_AllNestedFeaturesUnavailable_Patch {
             static bool Prefix(ref bool __result, UnitDescriptor unit, LevelUpState state, FeatureSelectionState selectionState, BlueprintFeature feature) {
                 IFeatureSelection selection = (feature as IFeatureSelection);
@@ -17,7 +17,7 @@ namespace TabletopTweaks.Bugfixes.UI {
                 return false;
             }
         }
-        [HarmonyPatch(typeof(FeatureSelectionExtensions), "CanSelectAny", new[] { typeof(IFeatureSelection), typeof(UnitDescriptor), typeof(LevelUpState), typeof(FeatureSelectionState) })]
+        //[HarmonyPatch(typeof(FeatureSelectionExtensions), "CanSelectAny", new[] { typeof(IFeatureSelection), typeof(UnitDescriptor), typeof(LevelUpState), typeof(FeatureSelectionState) })]
         static class FeatureSelectionExtensions_CanSelectAny_Patch {
             static bool Prefix(ref bool __result, IFeatureSelection selection, UnitDescriptor unit, LevelUpState state, FeatureSelectionState selectionState) {
                 FeatureSelectionState newSelectionState = new FeatureSelectionState(selectionState, selectionState.Source, selection, 0, 0);